Refactor: stage doc-engine extraction under src/doc-engine
