+++ doc/General-NL-Skeletons.md
@@ -1,0 +1,16 @@
+# General NL Skeletons
+
+## Structural Change Skeleton
+1. Goal and scope
+2. Current state
+3. Target state
+4. Invariants/constraints
+5. Step plan
+6. Validation checks
+
+## Behavior Change Skeleton
+1. User-visible behavior now
+2. User-visible behavior target
+3. Internal mechanism changes
+4. Compatibility notes
+5. Tests and verification
