+++ doc/NL-First-Workflow.md
@@ -1,0 +1,7 @@
+# NL-First Workflow
+
+## Workflow
+1. Capture/update a concise NL skeleton for the change before editing code.
+2. Keep NL step ordering aligned with implementation commits.
+3. Preserve constraints/invariants in NL and verify in tests/build.
+4. Update architecture docs when structural boundaries change.
